put spinner generation into helper function
